fix(src): Refactor in folders of src for apply clean arquitecture
